move local impl of FileSystem to separate mod
